Add power commands to SystemBridge
