+++ .coveragerc
@@ -1278,6 +1278,7 @@ omit =
     homeassistant/components/system_bridge/coordinator.py
     homeassistant/components/system_bridge/media_source.py
     homeassistant/components/system_bridge/sensor.py
+    homeassistant/components/system_bridge/services.py
     homeassistant/components/systemmonitor/sensor.py
     homeassistant/components/tado/__init__.py
     homeassistant/components/tado/binary_sensor.py
